Upload and download of config
can upload and download the config.
theres currently no protections for checking if the json is the right structure
and need to look into code injecting from this, like if theres ways to detect it, and then need to hash the information.
but for now the most basic functionality is working